feat: makes asdf-dotnet list all versions and improves performance
This commit completly removes the depency on dotnet releases json
maddness and instead uses a versions.txt file that gets updated on our
own repo.

This makes it almost instantly to list all version as we only need to
downloada simple txt file instead of waiting for hundreds of requests
and processing json with bash.